requirements.txt
alpaca-py==0.39.0
annotated-types==0.7.0
backtrader==1.9.78.123
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
fonttools==4.56.0
frozendict==2.4.6
idna==3.10
importlib_metadata==8.6.1
influxdb-client==1.48.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.1.0
multitasking==0.0.11
narwhals==1.29.1
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
reactivex==4.0.4
requests==2.32.3
retrying==1.3.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
soupsieve==2.6
sseclient-py==1.8.0
svgwrite==1.4.3
threadpoolctl==3.5.0
Tree==0.2.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
yfinance==0.2.54
zipp==3.21.0
